backend/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "smartcart-api"
version = "0.1.0"
description = "SmartCart AI — FastAPI backend for scan, optimise, and Gemini integration"
requires-python = ">=3.11,<3.14"
dependencies = [
    "fastapi",
    "uvicorn",
    "httpx",
    "python-dotenv",
    "google-generativeai",
    "supabase",
    "groq",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[tool.setuptools]
py-modules = ["main"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::FutureWarning",
]
